Move all rust-server XML examples to one sample
Meaning that the other can test our behaviour in the absence of XML.
