+++ samples/server/petstore/rust-server/output/openapi-v3/src/mimetypes.rs
@@ -4,6 +4,10 @@ pub mod responses {
     use hyper::mime::*;
 
     // The macro is called per-operation to beat the recursion limit
+    /// Create Mime objects for the response content types for XmlPost
+    lazy_static! {
+        pub static ref XML_POST_OK: Mime = "application/xml".parse().unwrap();
+    }
 
 }
