+++ samples/server/petstore/rust-server/output/openapi-v3/src/server/mod.rs
@@ -167,10 +167,21 @@ where
                                             Ok(rsp) => match rsp {
                                                 XmlPostResponse::OK
 
+                                                    (body)
+
 
                                                 => {
                                                     response.set_status(StatusCode::try_from(201).unwrap());
 
+                                                    response.headers_mut().set(ContentType(mimetypes::responses::XML_POST_OK.clone()));
+
+
+                                                    let mut namespaces = BTreeMap::new();
+                                                    // An empty string is used to indicate a global namespace in xmltree.
+                                                    namespaces.insert("".to_string(), models::namespaces::XMLOBJECT.clone());
+                                                    let body = serde_xml_rs::to_string_with_namespaces(&body, namespaces).expect("impossible to fail to serialize");
+
+                                                    response.set_body(body);
                                                 },
                                             },
                                             Err(_) => {
