+++ samples/server/petstore/rust-server/output/petstore-with-fake-endpoints-models-for-testing/src/models.rs
@@ -70,77 +70,6 @@ impl Animal {
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct AnimalFarm(Vec<Animal>);
-
-impl ::std::convert::From<Vec<Animal>> for AnimalFarm {
-    fn from(x: Vec<Animal>) -> Self {
-        AnimalFarm(x)
-    }
-}
-
-impl ::std::convert::From<AnimalFarm> for Vec<Animal> {
-    fn from(x: AnimalFarm) -> Self {
-        x.0
-    }
-}
-
-impl ::std::iter::FromIterator<Animal> for AnimalFarm {
-    fn from_iter<U: IntoIterator<Item=Animal>>(u: U) -> Self {
-        AnimalFarm(Vec::<Animal>::from_iter(u))
-    }
-}
-
-impl ::std::iter::IntoIterator for AnimalFarm {
-    type Item = Animal;
-    type IntoIter = ::std::vec::IntoIter<Animal>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.0.into_iter()
-    }
-}
-
-impl<'a> ::std::iter::IntoIterator for &'a AnimalFarm {
-    type Item = &'a Animal;
-    type IntoIter = ::std::slice::Iter<'a, Animal>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        (&self.0).into_iter()
-    }
-}
-
-impl<'a> ::std::iter::IntoIterator for &'a mut AnimalFarm {
-    type Item = &'a mut Animal;
-    type IntoIter = ::std::slice::IterMut<'a, Animal>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        (&mut self.0).into_iter()
-    }
-}
-
-impl ::std::ops::Deref for AnimalFarm {
-    type Target = Vec<Animal>;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl ::std::ops::DerefMut for AnimalFarm {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-
-impl AnimalFarm {
-    /// Helper function to allow us to convert this model to an XML string.
-    /// Will panic if serialisation fails.
-    #[allow(dead_code, non_snake_case)]
-    pub(crate) fn to_xml(&self) -> String {
-        serde_xml_rs::to_string(&self).expect("impossible to fail to serialize")
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct ApiResponse {
     #[serde(rename = "code")]
     #[serde(skip_serializing_if="Option::is_none")]
@@ -288,7 +217,7 @@ pub struct Capitalization {
     #[serde(skip_serializing_if="Option::is_none")]
     pub sca_eth_flow_points: Option<String>,
 
-    /// Name of the pet
+    /// Name of the pet 
     #[serde(rename = "ATT_NAME")]
     #[serde(skip_serializing_if="Option::is_none")]
     pub att_name: Option<String>,
@@ -511,7 +440,7 @@ impl EnumArrays {
 #[allow(non_camel_case_types)]
 #[repr(C)]
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize, Eq, Ord)]
-pub enum EnumClass {
+pub enum EnumClass { 
     #[serde(rename = "_abc")]
     _ABC,
     #[serde(rename = "-efg")]
@@ -522,7 +451,7 @@ pub enum EnumClass {
 
 impl ::std::fmt::Display for EnumClass {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
-        match *self {
+        match *self { 
             EnumClass::_ABC => write!(f, "{}", "_abc"),
             EnumClass::_EFG => write!(f, "{}", "-efg"),
             EnumClass::_XYZ_ => write!(f, "{}", "(xyz)"),
@@ -1065,7 +994,7 @@ impl OuterComposite {
 #[allow(non_camel_case_types)]
 #[repr(C)]
 #[derive(Debug, Clone, Copy, PartialEq, PartialOrd, Serialize, Deserialize, Eq, Ord)]
-pub enum OuterEnum {
+pub enum OuterEnum { 
     #[serde(rename = "placed")]
     PLACED,
     #[serde(rename = "approved")]
@@ -1076,7 +1005,7 @@ pub enum OuterEnum {
 
 impl ::std::fmt::Display for OuterEnum {
     fn fmt(&self, f: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
-        match *self {
+        match *self { 
             OuterEnum::PLACED => write!(f, "{}", "placed"),
             OuterEnum::APPROVED => write!(f, "{}", "approved"),
             OuterEnum::DELIVERED => write!(f, "{}", "delivered"),
