+++ samples/server/petstore/rust-server/output/rust-server-test/examples/server_lib/server.rs
@@ -15,8 +15,7 @@ use rust_server_test::{Api, ApiError,
                       DummyPutResponse,
                       FileResponseGetResponse,
                       HtmlPostResponse,
-                      RawJsonGetResponse,
-                      XmlPostResponse
+                      RawJsonGetResponse
 };
 use rust_server_test::models;
 
@@ -65,13 +64,6 @@ impl<C> Api<C> for Server<C> where C: Has<XSpanIdString>{
     fn raw_json_get(&self, context: &C) -> Box<Future<Item=RawJsonGetResponse, Error=ApiError>> {
         let context = context.clone();
         println!("raw_json_get() - X-Span-ID: {:?}", context.get().0.clone());
-        Box::new(futures::failed("Generic failure".into()))
-    }
-
-    /// Test XML
-    fn xml_post(&self, xml_object: models::XmlObject, context: &C) -> Box<Future<Item=XmlPostResponse, Error=ApiError>> {
-        let context = context.clone();
-        println!("xml_post({:?}) - X-Span-ID: {:?}", xml_object, context.get().0.clone());
         Box::new(futures::failed("Generic failure".into()))
     }
